pyproject.toml
[tool.poetry]
name = "reid"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
norfair = { extras = ["metrics", "video"], version = "^2.2.0" }
requests = "^2.32.3"
tqdm = "^4.67.1"
pyyaml = "^6.0.2"
seaborn = "^0.13.2"
numpy = "1.23.5"
protobuf = "4.21.2"
tensorboard = "^2.18.0"
ipython = "^8.31.0"
psutil = "^6.1.1"
torch = { version = "^2.0.1+cu118", source = "torch118" }
torchvision = { version = "^0.15.2+cu118", source = "torch118" }
insightface = "^0.7.3"
onnxruntime-gpu = "1.18.0"
opencv-python = "^4.11.0.86"
shapely = "^2.0.6"
torchreid = "^0.2.5"
gdown = "^5.2.0"
texttable = "^1.7.0"
ultralytics = "^8.3.70"
loguru = "^0.7.3"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
mlflow = "^2.20.2"
gitpython = "^3.1.44"
boto3 = "^1.36.22"
tritonclient = "^2.54.0"

[[tool.poetry.source]]
name = "torch118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[tool.poetry.group.test.dependencies]
ruff = "^0.9.4"
mypy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
